Analytics now only active after user gives consent
