+++ lib/types.ts
@@ -88,4 +88,6 @@ export interface EventData {
   eventName: 'page_view' | 'cta_click' | 'page_leave' | 'page_exit'
   pageId: string
   ctaId?: string
+  isMobile?: boolean
+  language?: string
 }
