+++ pages/_app.tsx
@@ -34,21 +34,22 @@ import 'styles/notion.css'
 // import 'prismjs/components/prism-typescript'
 // import 'prismjs/components/prism-bash'
 
-import { CookieBanner } from '@palmabit/react-cookie-law'
-import { bannerStyles } from '../styles/banner'
+// import { CookieBanner } from '@palmabit/react-cookie-law'
+// import { bannerStyles } from '../styles/banner'
 
 export default function App({ Component, pageProps }) {
-  return (
-    <>
-      <Component {...pageProps} />
-      <CookieBanner
-        message='Wir verwenden kleine Kekse - auch genannt "Cookies". Willst du die auch?'
-        wholeDomain={true}
-        onAccept={() => {}}
-        acceptButtonText='Ja, klar!'
-        managePreferencesButtonText='Einstellungen'
-        styles={bannerStyles}
-      />
-    </>
-  )
+  // return (
+  //   <>
+  //     <Component {...pageProps} />
+  //     <CookieBanner
+  //       message='Wir verwenden kleine Kekse - auch genannt "Cookies". Willst du die auch?'
+  //       wholeDomain={true}
+  //       onAccept={() => {}}
+  //       acceptButtonText='Ja, klar!'
+  //       managePreferencesButtonText='Einstellungen'
+  //       styles={bannerStyles}
+  //     />
+  //   </>
+  // )
+  return <Component {...pageProps} />
 }
